requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
pillow
groq
